+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v2.0.1
+
+- Fix config editor showing same projections multiple times
+- Fix swap coordinates for WGS84
+
 # v2.0.0
 
 - Updates @vcmap/core and @vcmap/ui to version 6.x